CREATE repository from VS project
A project with name HelloWorld was created using a C# console
application template in visual studio, with the directory option
checked.  This is an attempt to create a new git repository on
github.com from this project.  In the github for windows program,
clicked on + in upper right hand corner, then with "Create" selected,
added a repository name of HelloWorld, and a path that is the parent
path for the HelloWorld directory created by visual studio.